refactor(racing-car): 비즈니스, ui, utils 로직 분리 및 변수명 수정
