+++ src/main/java/racingcar/RacingCarMain.java
@@ -1,22 +1,20 @@
 package racingcar;
 
-import java.util.List;
-import java.util.Scanner;
-
 import racingcar.domain.Car;
 import racingcar.service.RacingService;
-import racingcar.ui.RacingCarInput;
 import racingcar.ui.RacingCarOutput;
+import racingcar.utils.CarsGenerator;
+import racingcar.utils.RoundNumberGenerator;
+
+import java.util.List;
 
 public class RacingCarMain {
-	public static final Scanner sc = new Scanner(System.in);
-
-	public static void main(String[] args) {
-		List<Car> cars = RacingCarInput.carNameInput();
-		int round = RacingCarInput.roundInput();
-		RacingService racingService = new RacingService(cars);
-		RacingCarOutput.printRoundResultMessage();
-		racingService.race(round);
-		RacingCarOutput.printWinners(racingService.findWinners());
-	}
+    public static void main(String[] args) {
+        List<Car> cars = CarsGenerator.carNameInput();
+        int round = RoundNumberGenerator.roundInput();
+        RacingService racingService = new RacingService(cars);
+        RacingCarOutput.printRoundResultMessage();
+        racingService.race(round);
+        RacingCarOutput.printWinners(racingService.findWinners());
+    }
 }
