+++ src/main/java/racingcar/utils/CarsGenerator.java
@@ -1,24 +1,35 @@
 package racingcar.utils;
 
+import racingcar.domain.Car;
+import racingcar.ui.RacingCarInput;
+import racingcar.validator.CarNameValidator;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import racingcar.domain.Car;
-import racingcar.validator.CarNameValidator;
-
 public class CarsGenerator {
-	private static final String NAME_DELIMITER = ",";
+    private static final String NAME_DELIMITER = ",";
 
-	private static List<String> separateCarNames(String userInput) {
-		return Arrays.asList(userInput.split(NAME_DELIMITER, -1));
-	}
+    public static List<Car> carNameInput() {
+        String userInput = RacingCarInput.userCarNameInput();
+        try {
+            return CarsGenerator.generateCars(userInput);
+        } catch (IllegalArgumentException e) {
+            System.out.println(e.getMessage());
+            return carNameInput();
+        }
+    }
 
-	public static List<Car> generateCars(String userInput) {
-		List<String> carNames = separateCarNames(userInput);
-		carNames.forEach(CarNameValidator::validate);
-		return carNames.stream()
-			.map(Car::new)
-			.collect(Collectors.toList());
-	}
+    public static List<Car> generateCars(String userInput) {
+        List<String> carNames = separateCarNames(userInput);
+        carNames.forEach(CarNameValidator::validate);
+        return carNames.stream()
+                .map(Car::new)
+                .collect(Collectors.toList());
+    }
+
+    private static List<String> separateCarNames(String userInput) {
+        return Arrays.asList(userInput.split(NAME_DELIMITER, -1));
+    }
 }
