+++ src/main/java/racingcar/utils/RandomIntegerGenerator.java
@@ -1,7 +1,6 @@
 package racingcar.utils;
 
 public class RandomIntegerGenerator {
-
 	public static int random(int start, int end) {
 		return (int)(Math.random() * (end - start)) + start;
 	}
